refactor(watch): extract shared Stat/SimpleBars/LoadingCard
Move three locally-defined helpers from SpaceDetail.tsx into reusable
modules under frontend/src/watch/components/ so the upcoming Feedback
page can share them without duplication.

SimpleBars gains an optional colorClass prop (default 'bg-blue-500')
so callers can render green / red bars without changing existing
SpaceDetail behavior.
