+++ frontend/src/watch/components/Stat.tsx
@@ -1,0 +1,10 @@
+import { Card } from '@/components/ui/card'
+
+export function Stat({ label, value }: { label: string; value: React.ReactNode }) {
+  return (
+    <Card className="p-4">
+      <p className="text-xs uppercase text-muted">{label}</p>
+      <p className="mt-1 text-2xl font-semibold">{value}</p>
+    </Card>
+  )
+}
